fix: avoid data race in conventionalcommits parser
The conventionalcommits parser is not concurrency safe, since it stores the
input in the machine's state:
https://github.com/leodido/go-conventionalcommits/blob/v0.9.0/parser/machine.go#L131

So we need to create a machine on a per-request basis at minimum,
instead of per-server.

This also fixes a data race around a user config when testing custom types.

Now, all tests passes with `go test -race ./...`.